# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThetaFrame API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weekly-frames:
    get:
      operationId: listWeeklyFrames
      tags: [weekly-frames]
      summary: List all weekly frames for the current user
      responses:
        "200":
          description: List of weekly frames
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyFrame"
    post:
      operationId: createWeeklyFrame
      tags: [weekly-frames]
      summary: Create or update a weekly frame (weekStart in body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeeklyFrameBody"
      responses:
        "200":
          description: Created or updated weekly frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyFrame"
  /weekly-frames/{weekStart}:
    get:
      operationId: getWeeklyFrame
      tags: [weekly-frames]
      summary: Get weekly frame for a week
      parameters:
        - name: weekStart
          in: path
          required: true
          schema:
            type: string
            description: Week start date in YYYY-MM-DD format (Monday)
      responses:
        "200":
          description: Weekly frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyFrame"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertWeeklyFrame
      tags: [weekly-frames]
      summary: Create or update a weekly frame
      parameters:
        - name: weekStart
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertWeeklyFrameBody"
      responses:
        "200":
          description: Upserted weekly frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyFrame"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    WeeklyStep:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        emoji:
          type: ["string", "null"]
      required:
        - id
        - text
    WeeklyFrame:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        weekStart:
          type: string
          description: Week start date YYYY-MM-DD (Monday)
        theme:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        nonNegotiables:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        recoveryPlan:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - weekStart
        - steps
        - nonNegotiables
    CreateWeeklyFrameBody:
      type: object
      properties:
        weekStart:
          type: string
          description: Week start date YYYY-MM-DD (Monday)
        theme:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        nonNegotiables:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        recoveryPlan:
          type: ["string", "null"]
      required:
        - weekStart
        - steps
        - nonNegotiables
    UpsertWeeklyFrameBody:
      type: object
      properties:
        theme:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        nonNegotiables:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyStep"
        recoveryPlan:
          type: ["string", "null"]
      required:
        - steps
        - nonNegotiables
...